+++ Api/NFLApi/Entities/Tables/MatchResult.cs
@@ -4,8 +4,8 @@ public class MatchResult
 {
     public Match Match { get; set; }
     public int MatchId { get; set; }
-    public int AwayTeamScore { get; set; }
-    public int HomeTeamScore { get; set; }
+    public short AwayTeamScore { get; set; }
+    public short HomeTeamScore { get; set; }
     public Team MatchWinner { get; set; }
     public string MatchWinnerId { get; set; }
     public Team MatchLoser { get; set; }
